1. 将按键按下计数，由静态8位变量，改为16位变量计数

+++ ABFM03/Core/ZJ_Drivers/ZJ_Src/02_Key_Process.c
@@ -539,9 +539,9 @@ void Dealwith_testmode_level_dec_long(void) //
 **************************************************************************************/
 static void get_key_msg(struct_KeyInfo* pInfo)
 {
-  static uint8_t power_key_press_cnt;
-  static uint8_t level_add_key_press_cnt;
-  static uint8_t level_dec_key_press_cnt;
+  static uint16_t power_key_press_cnt;
+  static uint16_t level_add_key_press_cnt;
+  static uint16_t level_dec_key_press_cnt;
 
   switch(pInfo->CurKey)
   {
